adding ability for tabbed search results
