+++ packages/react-heartwood-components/src/components/Tabs/Tabs.js
@@ -19,7 +19,7 @@ export type Props = {
 	isTruncatable?: boolean,
 
 	/** Optional class to add to the tab group */
-	className?: String
+	className?: string
 }
 
 //TODO properly define State type
